App/Gui: fix GeoFeatureGroupExtension when Gui is not loaded
Normally App::GeoFeatureGroupExtension relies on
Gui::ViewProviderGeoFeatureGroupExtension to provide which children are
climed to implement getSubObjects() for shape export (e.g.
Part.getShape(), and STEP export). And it does not work if Gui module
is not loaded.

This patch provides a partial fix. It is only partial, because in the
current architecture, there is no way to know which children are climed
without loading Gui.